ga990xa_ud3: some fixes for upstream changes

+++ src/mainboard/gigabyte/ga990xa_ud3/rd890_cfg.c
@@ -104,12 +104,12 @@ static void nb_platform_config(device_t nb_dev, AMD_NB_CONFIG *NbConfigPtr)
  *
  * prototype AGESA_STATUS (*CALLOUT_ENTRY) (UINT32 Param1, UINTN Param2, VOID* ConfigPtr);
  *
- * @param[in] u32 func               Northbridge CIMx CallBackId
- * @param[in] u32 data               Northbridge Input Data.
- * @param[in] AMD_NB_CONFIG *config  Northbridge configuration structure pointer.
+ * @param[in] func     Northbridge CIMx CallBackId
+ * @param[in] data     Northbridge Input Data.
+ * @param[in] *config  Northbridge configuration structure pointer.
  *
  */
-static u32 rd890_callout_entry(u32 func, u32 data, void *config)
+static u32 rd890_callout_entry(u32 func, uintptr_t data, void *config)
 {
 	u32 ret = 0;
 #ifndef __PRE_RAM__
